Make references to codegen not require installation
`make codegen` now builds a new version of the codegen tool before trying
to run it. The command also now looks to a binary in the code-generation
directory instead of requiring installation.

This also adds a `go fmt` command to the code generation to remove formatting
diffs from the bindata generation stage.