Fix fetching cloudformationtemplate object from k8s
Currently, when fetching cloudformationtemplate object from Kubernetes, we
fail to fetch the object from a specific namespace, even if the namespace
is specified in the spec for an AWS service object like S3Bucket, etc., and
coerce to using the 'default' namespace, which may not contain the cloudformation
template objects.

This pull request allows fetching cloudformation temaplate objects from
specified namespaces.